#9 Set up main.py and timeseries.py to plot moving average in Tab1.
