+++ bokeh_app/main.py
@@ -10,16 +10,15 @@ from bokeh.models.widgets import Tabs
 
 
 # Each tab is drawn by one script
-# from scripts.plotting import plotting_tab
-from scripts.histogram import histogram_tab
+from scripts.timeseries import timeseries_tab
+# from scripts.histogram import histogram_tab
 
 # Read data into dataframes
 # Import data.
-ts = pd.read_csv("./bokeh_app/data/test_timeseries.csv", skiprows=1, delimiter=",", names=['time', 'y1', 'y2', 'y3'])
-ts.set_index('time', inplace=True, drop=True)
+ts = pd.read_csv("./bokeh_app/data/test_timeseries_copy.csv", skiprows=1, delimiter=",", names=['time', 'y1', 'y2', 'y3'])
 
 # Create each of the tabs
-tab1 = histogram_tab(ts)
+tab1 = timeseries_tab(ts)
 
 # Put all the tabs into one application
 tabs = Tabs(tabs=[tab1])
